pipeline/requirements.txt
# Data sources
yfinance>=1.2,<2
fredapi>=0.5,<1

# GCP services
google-cloud-firestore>=2.19,<3
google-cloud-storage>=2.19,<3
